ranamed "dialogs" package to "pcdialogs" for continutity
