+++ src/dmtools/GUI/entityguicomponents/pcdialogs/EditPcDialog.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dmtools.GUI.entityguicomponents.dialogs;
+package dmtools.GUI.entityguicomponents.pcdialogs;
 
-import dmtools.GUI.entityguicomponents.dialogs.panels.CreateACandHP;
-import dmtools.GUI.entityguicomponents.dialogs.panels.CreateSelectedSkills;
-import dmtools.GUI.entityguicomponents.dialogs.panels.CreateStatBlockPanel;
 import dmtools.GUI.entityguicomponents.comboboxes.EntityComboBox;
+import dmtools.GUI.entityguicomponents.pcdialogs.panels.DisplayACandHP;
+import dmtools.GUI.entityguicomponents.pcdialogs.panels.DisplaySelectedSkills;
+import dmtools.GUI.entityguicomponents.pcdialogs.panels.DisplayStatBlock;
 import dmtools.filehandling.FileHandler;
 import dmtools.game.entities.PC;
 import dmtools.game.entities.characteristics.Alignment;
@@ -35,12 +35,13 @@ import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
+import static javax.swing.WindowConstants.DISPOSE_ON_CLOSE;
 
 /**
  *
  * @author A3
  */
-public class CreatePcDialog extends JDialog implements PropertyChangeListener {
+public class EditPcDialog extends JDialog implements PropertyChangeListener {
 
     private static JLabel cNameLabel;
     private static JLabel pNameLabel;
@@ -56,21 +57,23 @@ public class CreatePcDialog extends JDialog implements PropertyChangeListener {
     private static JComboBox alignmentCB;
     private static JComboBox classCB;
     private static EntityComboBox levelBox;
-    private static CreateStatBlockPanel statBlockPanel;
-    private static CreateSelectedSkills selectedSkillList;
-    private static CreateACandHP acAndHP;
+    private static DisplayStatBlock statBlockPanel;
+    private static DisplaySelectedSkills selectedSkillList;
+    private static DisplayACandHP acAndHP;
 
     private JOptionPane optionPane;
 
-    private PC createdPC;
+    private PC updatedPC;
+    private PC toEditPC;
     private boolean shouldSave;
 
-    public CreatePcDialog(Frame frame, boolean shouldSave) {
+    public EditPcDialog(Frame frame, PC pc) {
         super(frame, true);
-        this.shouldSave = shouldSave;
+        this.toEditPC = pc;
+        this.shouldSave = true;
 
-        setTitle("Create a new Player Character");
-        String createButtonTxt = "Create";
+        setTitle(pc.getName());
+        String createButtonTxt = "Save";
         Object[] options = {createButtonTxt};
 
         optionPane = new JOptionPane(getInputFields(),
@@ -107,14 +110,13 @@ public class CreatePcDialog extends JDialog implements PropertyChangeListener {
              */
             List<String> errors = checkErrors();
             if (errors == null) {
-                //create the pc here
-                this.createdPC = createPC();
+                //save the pc here
+                this.updatedPC = getPcWithChanges();
                 if (shouldSave) {
                     try {
-                        FileHandler.saveFromInstance(createdPC, 
+                        FileHandler.saveFromInstance(updatedPC, 
                                 FileHandler.PC_FILE);
                     } catch (IOException pcFileHandlerException) {
-                        System.out.println(pcFileHandlerException.getLocalizedMessage());
                     }
                 }
                 dispose();
@@ -128,7 +130,7 @@ public class CreatePcDialog extends JDialog implements PropertyChangeListener {
         }
     }
 
-    private PC createPC() {
+    private PC getPcWithChanges() {
         String characterName = cNameTF.getText();
         String playerName = pNameTF.getText();
         Race race = Race.stringValueOf(raceCB.getSelectedItem().toString());
@@ -151,45 +153,54 @@ public class CreatePcDialog extends JDialog implements PropertyChangeListener {
 
     }
     
-    public PC getCreatedPC() {
-        return createdPC;
+    public PC getUpdatedPC() {
+        return updatedPC;
     }
 
-    private static JComponent[] getInputFields() {
+    private JComponent[] getInputFields() {
         //Character name
         cNameLabel = new JLabel("Character Name");
-        cNameTF = new JTextField();
+        cNameTF = new JTextField(toEditPC.getName());
+        cNameTF.setEnabled(false);
+        
 
         //Player name
         pNameLabel = new JLabel("Player Name");
-        pNameTF = new JTextField();
+        pNameTF = new JTextField(toEditPC.getPlayerName());
+        pNameTF.setEnabled(false);
 
         //Race
         raceLabel = new JLabel("Race");
         raceCB = new EntityComboBox(EntityComboBox.RACE_BOX);
+        raceCB.setSelectedItem(toEditPC.getRace().toString());
+        raceCB.setEnabled(false);
 
         //Alignment
         alignmentLabel = new JLabel("Alignment");
         alignmentCB = new EntityComboBox(EntityComboBox.ALIGNMENT_BOX);
+        alignmentCB.setSelectedItem(toEditPC.getAlignment().toString());
 
         //Playable Class
         classLabel = new JLabel("Class");
         classCB = new EntityComboBox(EntityComboBox.PLAYABLE_CLASS_BOX);
+        classCB.setSelectedItem(toEditPC.getPlayableClass().toString());
+        classCB.setEnabled(false);
 
         //Character Level
         cLevelLabel = new JLabel("Character Level");
         levelBox = new EntityComboBox(EntityComboBox.CHARACTER_LEVEL_BOX);
+        levelBox.setSelectedIndex(toEditPC.getCharacterLevel() - 1);
 
         //StatBlock
-        statBlockPanel = new CreateStatBlockPanel();
+        statBlockPanel = new DisplayStatBlock(toEditPC);
         statBlockPanel.setFocusable(false);
 
         //Selected Skills
-        selectedSkillList = new CreateSelectedSkills();
+        selectedSkillList = new DisplaySelectedSkills(toEditPC);
         selectedSkillList.setFocusTraversalKeysEnabled(false);
 
         //AC, Max HP, Current HP
-        acAndHP = new CreateACandHP();
+        acAndHP = new DisplayACandHP(toEditPC);
 
         //AC&HP, Statblock and Selected Skills in a panel
         JPanel AcHpStatAndSkillPanel = new JPanel();
@@ -306,18 +317,20 @@ public class CreatePcDialog extends JDialog implements PropertyChangeListener {
             errorList.add("-HP cannot be blank");
             acAndHP.highlight(true, "HP");
         }
+        
+        //selected skills disabled
 
-        //Checks Selected Skills > 0
-        if (selectedSkillList.getSelectedSkills().size() < 1) {
-            errorList.add("-At least one skill must be selected");
-            selectedSkillList.highlight(true);
-            /*
-            *THIS IS WHERE YOU WILL PUT THE INFO FOR SKILLS SELECTED != NUMBER
-            *OF POSSIBLE CLASS SKILLS
-             */
-        } else {
-            selectedSkillList.highlight(false);
-        }
+//        //Checks Selected Skills > 0
+//        if (selectedSkillList.getSelectedSkills().size() < 1) {
+//            errorList.add("-At least one skill must be selected");
+//            selectedSkillList.highlight(true);
+//            /*
+//            *THIS IS WHERE YOU WILL PUT THE INFO FOR SKILLS SELECTED != NUMBER
+//            *OF POSSIBLE CLASS SKILLS
+//             */
+//        } else {
+//            selectedSkillList.highlight(false);
+//        }
 
         //Stats
         statBlockPanel.resetHighlight();
